Add persisted hidden state per product and handle expected error states
